(fixed 64bit compiler warnings)
svn-id: https://kwave.svn.sourceforge.net/svnroot/kwave/trunk@2215 bbf32e5a-961d-0410-9e10-91f32741c31e

+++ libkwave/SwapFile.cpp
@@ -89,7 +89,7 @@ bool SwapFile::allocate(size_t size, const QString &filename)
 #endif/* HAVE_MKSTEMP */
 
     m_file.seek(round_up(size, m_pagesize));
-    if (m_file.pos() + 1 < size) {
+    if ((qint64)(m_file.pos() + 1) < (qint64)size) {
 	qWarning("SwapFile::allocate(%d MB) failed, DISK FULL ?",
 	         (unsigned int)(size >> 20));
 	m_size = 0;
